Add the `shellfunction` decorator
Experimental feature to allow running a shell command by decorating a
Python function as opposed to running a full-blown `ShellJob`. The
advantage is that it should have a lot less overhead and so be faster,
however, it can not be run on an arbitrary computer or sent to the
daemon.