Tests: add a test that submits `ShellJob` to the daemon
To make it easy to write the test, a new fixture `daemon` is added. This
fixture has a session scope and will launch the daemon for the current
default profile, which should be the profile being used to run the
tests.

The daemon is started by invoking `verdi daemon start-circus` as a sub
process through the `subprocess` module. This is necessary because the
code that launches the daemon backgrounds itself, so it cannot be called
from the interpreter that is running the tests itself.

At the end of the session the daemon should be automatically stopped,
again by invoking `verdi daemon stop` in a subprocess.